WWTExplorer3d/Healpix/HealpixTile.cs: add HiPS analytics
Add a query-string to HiPS tile URLs that will allow providers to get some
analytics on traffic sources. With the Windows client, we could set the
User-Agent, but that's not an option for the web-based client(s), and it will
keep things cleaner to use a single mechanism across platforms.